Avoid signed shift into sign bit at two callsites
(1 << 31) shifts a signed 1 into the sign bit, which is
implementation-defined in C++17 and undefined in earlier standards.
Other places in the codebase that need bit 31 (for example
auto_param.cpp:1875 and ui.h:89) use ((uint32_t)1 << 31). Match that
pattern in Output::writeDataToFile and the WaveTable copy macro.

+++ src/deluge/model/output.cpp
@@ -281,7 +281,7 @@ bool Output::writeDataToFile(Serializer& writer, Clip* clipForSavingOutputOnly, 
 				else {
 					clipCode = thisInstance->clip->indexForSaving;
 					if (thisInstance->clip->section == 255) {
-						clipCode |= (1 << 31);
+						clipCode |= ((uint32_t)1 << 31);
 					}
 				}
